FIX: Gradient anisotropic diffusion requires float or double precision image types, changed input and output image types from short to double

+++ src/GradientAD/GradientAD.cxx
@@ -7,8 +7,8 @@
 #include "itkImageFileReader.h"
 #include "itkImageFileWriter.h"
 
-using InputPixelType =  short;
-using OutputPixelType = short;
+using InputPixelType =  double;
+using OutputPixelType = double;
 constexpr int Dim = 3;
